+++ src/TitleDetails/ExpectedPiecesList/ExpectedPiecesList.js
@@ -1,9 +1,10 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
+import { PIECE_COLUMNS } from '../constants';
 import PiecesList from '../PiecesList';
 
-const visibleColumns = ['caption', 'format', 'receiptDate', 'request', 'selection'];
+const visibleColumns = [PIECE_COLUMNS.caption, 'format', PIECE_COLUMNS.receiptDate, 'request', 'selection'];
 
 const ExpectedPiecesList = ({ pieces, selectPiece }) => {
   return (
@@ -11,6 +12,7 @@ const ExpectedPiecesList = ({ pieces, selectPiece }) => {
       pieces={pieces}
       selectPiece={selectPiece}
       visibleColumns={visibleColumns}
+      sortedColumn={PIECE_COLUMNS.receiptDate}
     />
   );
 };