+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## (IN PROGRESS)
 
+* permission sets should avoid .all permissions. Refs UIREC-141.
 * The Delete button is displayed in the Add piece pop up (new pice). Refs UIREC-146.
 
 ## [1.4.0](https://github.com/folio-org/ui-receiving/tree/v1.4.0) (2021-06-17)